requirements.txt
asgiref==3.4.1
boto3==1.18.1
botocore==1.21.1
dj-database-url==0.5.0
Django==3.2.5
django-social-share==2.2.1
django-storages==1.11.1
django-tinymce4-lite==1.8.0
gunicorn==20.1.0
jmespath==0.10.0
jsmin==2.2.2
Pillow==8.3.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
